services：verify用Integer而服务端/排序器用Long，大identity客户端假阴性
问题：
- 证据：服务身份数字口径三处不一致——服务端isLegalServiceIdentity与
  BServiceInfos排序器均Long.parseLong，客户端verify用Integer.parseInt：
  (Integer.MAX,Long.MAX]的数字identity（如雪花id）服务端接受、排序
  支持，但registerService在verify处抛NumberFormatException被客户端
  自己拒绝。

修复：
- verify改Long.parseLong，三处以Long为唯一口径。

验证：
- 红绿：TestSmAgentMisc.testVerifyLongIdentity（Long.MAX/Integer.MAX+1
  通过、非法串拒绝）修复前FAILED（stash验证）、修复后通过；全量fast通过。

+++ ZezeJava/ZezeJava/src/main/java/Zeze/Services/ServiceManager/AbstractAgent.java
@@ -421,8 +421,10 @@ public abstract class AbstractAgent extends ReentrantLock implements Closeable {
 
 	protected static void verify(@NotNull String identity) {
 		if (!identity.startsWith("@") && !identity.startsWith("#")) {
+			// Long口径（FND4-63）：服务端isLegalServiceIdentity与BServiceInfos排序器均Long.parseLong，
+			// 原Integer.parseInt把(Integer.MAX,Long.MAX]的数字identity（如雪花id）在客户端误拒。
 			//noinspection ResultOfMethodCallIgnored
-			Integer.parseInt(identity);
+			Long.parseLong(identity);
 		}
 	}
